added a code that converts the ips to an int64 format

+++ scripts/preprocessing.py
@@ -25,21 +25,27 @@ def convert_ip_to_int(df):
     return df
 
 def merge_ip_country(fraud_data, ip_country_data):
-    # Sort by lower bound of IP range to ensure correct merge_asof behavior
+    # Convert IP columns to int64 for both fraud_data and ip_country_data
+    fraud_data['ip_address'] = fraud_data['ip_address'].astype(int)
+    ip_country_data['lower_bound_ip_address'] = ip_country_data['lower_bound_ip_address'].astype(int)
+    ip_country_data['upper_bound_ip_address'] = ip_country_data['upper_bound_ip_address'].astype(int)
+
+    # Sort the ip_country_data by 'lower_bound_ip_address'
     ip_country_data = ip_country_data.sort_values(by='lower_bound_ip_address')
 
-    # Merge based on IP address falling within the range [lower_bound_ip_address, upper_bound_ip_address]
+    # Perform the range-based merge using merge_asof
     fraud_data = pd.merge_asof(fraud_data.sort_values('ip_address'), 
                                ip_country_data[['lower_bound_ip_address', 'upper_bound_ip_address', 'country']],
                                left_on='ip_address', 
                                right_on='lower_bound_ip_address',
                                direction='backward')
 
-    # Filter to retain only rows where ip_address falls within the valid range
+    # Filter to retain only rows where the ip_address falls within the valid IP range
     fraud_data = fraud_data[(fraud_data['ip_address'] >= fraud_data['lower_bound_ip_address']) & 
                             (fraud_data['ip_address'] <= fraud_data['upper_bound_ip_address'])]
 
     return fraud_data
+
 
 def add_transaction_features(df):
     # Time between signup and purchase (recency feature)
